pyproject.toml
[tool.poetry]
name = "meta"
version = "0.1.0"
description = ""
authors = ["Faisal-Kushha <fkushha89@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.4"
black = "^22.12.0"
flake8 = "^6.0.0"
jwt = "^1.3.1"
python-jwt = "^4.0.0"
django-hashid-field = "^3.3.7"
hashids = "^1.3.1"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
